requirements.txt
blinker==1.4
cachelib==0.9.0
cachetools==5.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Cors==3.0.10
Flask-Mail==0.9.1
Flask-Session==0.4.0
google-api-core==2.8.0
google-api-python-client==2.48.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.1
gphotospy==0.2.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.0
passlib==1.7.4
Pillow==9.1.1
protobuf==3.20.1
py-bcrypt==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.9
Werkzeug==2.1.2
